elderease/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework.authtoken.views import ObtainAuthToken

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('app.urls')),
    path('api-token-auth/', ObtainAuthToken.as_view(), name='api_token_auth'),
]
