requirements.txt
Django>=4.2
djangorestframework>=3.20
djangorestframework-simplejwt>=6.0